PKI_Certificates/requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
Naked==0.1.31
pycryptodome==3.9.9
PyYAML==5.3.1
requests==2.25.0
shellescape==3.8.1
urllib3==1.26.2
